Rework settings service.
